requirements.txt
# StockPulse 360 Requirements
# Install these packages for local development

# Snowflake packages
snowflake-snowpark-python>=1.11.0
snowflake-connector-python>=3.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
streamlit>=1.29.0
plotly>=5.18.0

# Optional: For email alerts
# python-dotenv>=1.0.0

# Optional: For Slack/webhook alerts
# requests>=2.31.0
